add list commands. resolve deps not actually working

+++ setups/bash/pre-config.sh
@@ -7,17 +7,17 @@ set -euo pipefail
 #  mkdir -p "$HOME/.local/bin"
 #fi
 
-if [ -f "$HOME/.profile" ]; then
+if [ ! -L "$HOME/.profile" ] && [ ! -f "$HOME/.profile.bak" ]; then
   echo "[setup] backing up ${HOME}/.profile"
   mv "$HOME/.profile" "$HOME/.profile.bak"
 fi
 
-if [ -f "$HOME/.bashrc" ]; then
+if [ ! -L "$HOME/.bashrc" ] && [ ! -f "$HOME/.bashrc_bak" ]; then
   echo "[setup] backing up ${HOME}/.bashrc"
   mv "$HOME/.bashrc" "$HOME/.bashrc_bak"
 fi
 
-if [ -f "$HOME/.bash_logout" ]; then
+if [ ! -L "$HOME/.bash_logout" ] && [ ! -f "$HOME/.bash_logout.bak" ]; then
   echo "[setup] backing up ${HOME}/.bash_logout"
   mv "$HOME/.bash_logout" "$HOME/.bash_logout.bak"
 fi
